coding-assist-be/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
openai
python-dotenv
pytest
httpx
psycopg2-binary
boto3
PyJWT
passlib[bcrypt]
